Moving things around and removing polaris
