+++ packages/react-toast/src/hooks.ts
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import {Context} from './context';
+
+export const useToasts = () => React.useContext(Context);
