+++ packages/react-toast/src/types.ts
@@ -1,0 +1,7 @@
+export interface ToastDescription {
+  content: string;
+  error?: boolean;
+  dismissible?: boolean;
+  duration?: number;
+  onDismiss?(): void;
+}
